Forgot to update a newer version of IEDB Tool
The format of the experimentally validated epitope csv file has changed in IEDB. The previous tool which used the simple csv format doesn't work anymore. Additionally that format is also not available to download so I've just whipped up a very quick fix a while back. It's not what I would have liked, but for now it works. I need to make some time to actually fix fix this code. Apologies to anyone who tried to use this (if there are anybody ahahahahah).  (~w~)7